+++ test/fixtures/react-hook-invalid/src/pages/index.tsx
@@ -1,0 +1,6 @@
+import { useCallback } from "react"
+
+export default function Page() {
+  const callback = useCallback(() => {}, [Date.now()])
+  return <button onClick={callback}>Invalid</button>
+}